refactor(interpreter): split declaration and assignment tests
